prezi/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-pptx==0.6.23
# LLM Providers
anthropic>=0.18.1
openai>=1.12.0
# Research Providers
google-search-results>=2.4.2
requests>=2.31.0
# Core
pydantic>=2.6.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
matplotlib>=3.7.0
pillow>=10.0.0
aiofiles>=23.2.1
httpx>=0.26.0
# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
